backend/requirements.txt
pandas
scikit-learn
google-cloud-translate
mysql-connector-python
python-dotenv
openai